requirements.txt
openai-whisper
transformers
torch
sentencepiece
protobuf
